@@ -6,15 +6,16 @@ Console.OutputEncoding = System.Text.Encoding.UTF8;
 decimal precoInicial = 0;
 decimal precoPorHora = 0;
 
-Console.WriteLine("Seja bem vindo ao sistema de estacionamento!\n" +
-                  "Digite o preço inicial:");
+Console.Clear();
+Console.Write("Seja bem vindo ao sistema de estacionamento!\n" +
+                  "Digite o preço inicial: ");
 precoInicial = Convert.ToDecimal(Console.ReadLine());
 
-Console.WriteLine("Agora digite o preço por hora:");
+Console.Write("Agora digite o preço por hora: ");
 precoPorHora = Convert.ToDecimal(Console.ReadLine());
 
 // Instancia a classe Estacionamento, já com os valores obtidos anteriormente
-Estacionamento es = new Estacionamento(precoInicial, precoPorHora);
+Estacionamento es = new(precoInicial, precoPorHora);
 
 string opcao = string.Empty;
 bool exibirMenu = true;
@@ -23,23 +24,39 @@ bool exibirMenu = true;
 while (exibirMenu)
 {
     Console.Clear();
+
+    Console.WriteLine("""
+    
+    ██████╗░░█████╗░██████╗░██╗░░██╗██╗███╗░░██╗░██████╗░  ██████╗░██╗░█████╗░
+    ██╔══██╗██╔══██╗██╔══██╗██║░██╔╝██║████╗░██║██╔════╝░  ██╔══██╗██║██╔══██╗
+    ██████╔╝███████║██████╔╝█████═╝░██║██╔██╗██║██║░░██╗░  ██║░░██║██║██║░░██║
+    ██╔═══╝░██╔══██║██╔══██╗██╔═██╗░██║██║╚████║██║░░╚██╗  ██║░░██║██║██║░░██║
+    ██║░░░░░██║░░██║██║░░██║██║░╚██╗██║██║░╚███║╚██████╔╝  ██████╔╝██║╚█████╔╝
+    ╚═╝░░░░░╚═╝░░╚═╝╚═╝░░╚═╝╚═╝░░╚═╝╚═╝╚═╝░░╚══╝░╚═════╝░  ╚═════╝░╚═╝░╚════╝░
+    """+"\n");
+
     Console.WriteLine("Digite a sua opção:");
     Console.WriteLine("1 - Cadastrar veículo");
     Console.WriteLine("2 - Remover veículo");
     Console.WriteLine("3 - Listar veículos");
     Console.WriteLine("4 - Encerrar");
+    Console.Write("Digite uma das opções: ");
+    opcao = Console.ReadLine();
 
-    switch (Console.ReadLine())
+    switch (opcao)
     {
         case "1":
+            Console.Clear();
             es.AdicionarVeiculo();
             break;
 
         case "2":
+            Console.Clear();
             es.RemoverVeiculo();
             break;
 
         case "3":
+            Console.Clear();
             es.ListarVeiculos();
             break;
 
@@ -51,8 +68,8 @@ while (exibirMenu)
             Console.WriteLine("Opção inválida");
             break;
     }
-
-    Console.WriteLine("Pressione uma tecla para continuar");
+    
+    Console.Write("Pressione uma tecla para continuar: ");
     Console.ReadLine();
 }
 
